change stdout to transient temperature
HotSpot v6.0 dumps a temperature log to stdout, which can be used as
input for the next call to HotSpot. HotSpot v7.0 changed the stdout log
to simulation status messages. This commit reverts that change, so the
stdout log should now be a dump of the transient temperatures, just
HotSniper expects. Also, we cannot use the transient temperature output
file as this log excludes the temperatures of the heatsink.