fix: HotSpot outputs in kelvin instead of celsius
HotSpot v7.0 changed the output temperature unit from degrees celsius to
kelvin for transient temperatures file. However, this is not mentioned
in the changelog and the comment in that part of the code claims to
output in degrees celsius. HotSniper expects celsius, so this change of
unit is marked as a bug and fixed.